Add defensive code when returning server's client connect URL
We saw a case where a Windows VM would return 0.0.0.0 in the INFO
protocol. After debugging, it appeared that the VM had an issue
and a reboot solved it, but while in bad mode, the resolution of
the IP address would not return an error, but nothing was added
to the array and we ended-up adding at the last resort the 0.0.0.0
IP. This is a problem for Windows clients since they can't use
that IP to connect.
We could simply do nothing and the client would fail to reconnect
using this IP - which is fine -, but this is masking the real issue,
so deciding to panic when detecting this situation.

/cc @derekcollison @ColinSullivan1