[FIXED] Override of config file options with command line parameters
The idea is to call flag.Parse() twice. The first time - as it was
the case before - is setting the empty Options structure with all
the specified flag defaults and with values provided from the command
line.
If a config file is specified, we then apply to this options structure
the content of the config file. This is done with a new receiver
function for the Options structure. After that, we call flag.Parse() again
which will override the latest changes with any command line parameter.
We don't need MergeOptions() anymore since the above process takes
care of the proper override.

Resolves #574