[FIXED] Profiler use of WriteTimeout was preventing so trace collection
During a recent change related to port, the profiler http server
config was changed and WriteTimeout of 2 seconds was added.
This was breaking profiling requests that check that the server
WriteTimeout is not less than the request's duration (for instance
it would be impossible to collect a trace of 5 seconds with such
setting).