[FIXED] Removing fan-in fix
The commit 4d48a17ed582cb00f17f6069fa155725690d9567 causes unwanted
behavior in some deployments. Reverting this change in this branch.
More work on solving fan-in issues is being done in the master branch.

Users affected by fan-in issues since 1.2.0, can use version 1.4.1.
Users that were not but are now affected with the fan-in fix in
1.4.0/1.4.1 should use this revision or upgrade to 2.0.0 when GA..

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>