[CHANGED] MQTT: do not require bearer JWT
In operator mode, all connections require a JWT. For MQTT clients,
the JWT is passed as the password in the MQTT connect protocol.

Since MQTT clients do not participate in the NATS protocol: do not
receive an INFO from server and can't sign the nonce, a requirement
was made that MQTT clients use a JWT with bearer boolean set to true,
which then cause the server to skip the signature check.

I propose that we simply bypass the signature if the JWT has bearer
token set to true OR if the client is MQTT.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>