[FIXED] JetStream: possible races during concurrent account lookup
This could leave the registered account in s.accounts with a
jsAccount that was not the one registered in js.accounts map.

Also fixed the filestore flushLoop invocation to pass the channels
to avoid races that could leave that go routine running at the
end of a test.

Fixed source and mirror consumers routines that could too be
left running on server shutdown.

Added missing defer for client connection.Close() and some servers
shutdown.

Move a test that was not starting with TestJetStreamCluster out
of the jetstream_cluster_test.go file and into jetstream_test.go.

Fixed a test name that did not have proper TestJetStream prefix in
jetstream_leafnode_test.go.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>