Fix setting `InactivityThreshold` on existing durable consumer
Beforehand, the threshold and timer weren't resetting properly when
an existing durable consumer was updated with a new `InactivityThreshold`
so consumers wouldn't get cleaned up until the NATS Server was restarted.
This PR refactors that to use a common function both when the consumer
is set up and when it is updated, which effectively fixes the bug. It also
updates the unit tests to verify that it works both ways.