Prevent duplicate responses from internal response subscriptions
There can be an issue with leafnodes and service imports where there
might already exist a subscription that represents the same interest
as the service response import.

In order to avoid duplicate responses due to both `_R_.foo.>` and `_R_.foo.bar`
being registered, we supress adding an extra subscription and using the one
that is already present.

Signed-off-by: Waldemar Quevedo <wally@nats.io>