Cheery-picked: [FIXED] Increased MQTT test R/W timeout from 4 to 5s (TestMQTTSubPropagation) (#4517)
Tracing the connect (ack?) read times in `TestMQTTSubPropagation` showed
that they come in the 2-3s range during normal execution, and it appears
that they occasionally exceed the 4s timeout.

I am not sure exactly why MQTT CONNECT takes such a long time, but as
the name of the test suggests, perhaps it has to do with session
propagation in a cluster.