[IMPROVED] Routing: upgrade non solicited routes if present in config
When a server accepts a route, that route would not be marked as
solicited or configured while it should if it was present in the
configuration.

Prior to pooling, if server A pointed to server B and vice-versa,
one of the server will have the route as solicited and the server
that accepted the route would have upgraded it as solicited.
With pooling code, this was not always happening.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>