Improved time to select skip list and starting sequence number for deliver last by subject.
Previously we would need to move through all msg blocks and load all of the blocks into memory.
We optimized this regardless to flush the cache if we loaded it like other places in the code, but for this specifically we load subject's totals which is already in memory and for each matching subject we now simply load the last message for that subject and add the sequence number to the skip list.

This drastically improves consumers that want last per subject, like KV watchers and object store lists.

Signed-off-by: Derek Collison <derek@nats.io>