[IMPROVED] MQTT: do not JSON-encode retained message bodies (#4825)
#### BREAKING

This change is not backwards-compatible. It will break in a mixed
environment and/or when downgrading.

if a server runs this code, and the cluster/server is reverted to an
older version or there are servers in the cluster that are still running
older versions, the retained messages will fail to be loaded.

#### SUMMARY

We used to JSON-encode the payloads of retained messages when storing to
JetStream. This PR moves the JSON metadata into the NATS header, and
stores the message bytes unencoded in the NATS message body.