[FIXED] Fix unsupported MQTT characters error message
Previously ' ' and '.' were not supported for MQTT topics until #4243 brought in support for '.'. However the error message for unsupported characters appears to not have been updated. This simply updates the error message.

Had an example of a MQTT client trying to publish to `spBv1.0/Field Devices/Some Device/Some Topic` and the error message indicated that both '.' and ' ' were unsupported which was misleading as only ' ' is unsupported.