Added OpenTelemetry support
If the `Nats-Trace-Dest` header is not present, but `Traceparent` is
and its last token is `01`, then message tracing is triggered. This
also requires that the account be defined with a `trace_dest` subject
so that traces can be sent there.
Note that `Nats-Trace-Only` is not applicable for `Traceparent`.

Addition to PR #5014
Resolves #5052

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>