Add support for AsOfTime for get/direct msg requests
The motivation for this feature is to enable stream message gets
as of some starting time. This is to match the same behavior with
setting a start time when creating a consumer.

What this enables, combined with batches, is to read messages from
a stream without the need to use a consumer. This is not intended
for high-throughput use, but for querying and stream inspection.

The main driving need for this is the ability to inspect workqueue
streams that cannot  otherwise do due to the overlapping subject
constraint.

Signed-off-by: Byron Ruth <byron@nats.io>