[FIXED] Adjust JSZ account filtering when asking for stream details
Normally when making a JSZ request with an account filter, the server 
sends back a single account details.

However if stream info is requested AND  the server did not find the 
account in the filter, then the server would send back ALL accounts.

With this change, if an account filter is provided, the server will 
respond with one account detail or none.