NRG: Only reset election timeout on vote request if granting
With the current behaviour, a node that is behind but running in the
candidate state could cause nodes that are current to delay switching
to the candidate state themselves because they keep resetting the
election timer each time they receive a vote request.

Changing this to reset the election timeout only when granting means
that we will allow other possibly more up-to-date nodes to transition to
the candidate state faster.

This might explain why we've seen some unexpected delays in starting a
new election after a leader has gone away.

Signed-off-by: Neil Twigg <neil@nats.io>