Release the fs lock when determing next first sequence for a subject.
In cases where max per subject is utilized, once that condition is hit storing a new message involves writing the new one, flushing it, then determining the current first seq for the subject so we can remove it. Determing the first sequence for the subject could involve walking multiple blocks and loading them in as well, all while the filestore lock is being held causing delays for other access to the filestore, even for things like FastState().

Signed-off-by: Derek Collison <derek@nats.io>