NRG: Stepdown candidate when AE ahead of last log term
A node that has been isolated in the candidate state for a considerable
amount of time might have bumped its vote term up considerably due to a
number of failed election attempts, so when the network recovers, new
AEs would not trigger a stepdown. Instead, check the `pterm`, which is
the term of the last entry in our log, as this will not have ratcheted
up during the period of isolation.

Signed-off-by: Neil Twigg <neil@nats.io>