Improvements to handling snapshots for memory based wals.
With memory based wals we can not use snapshots on restarts, but we do use them while they are running.
However if a server becomes a leader with no snapshot it will be forced to stepdown when asked to catchup a follower.
So we now inherit a leaders snapshot.

Also when we tried to truncate on a mismatch, we needed to truncate the previous index, not current.
When we fail due to the previous entry being compacted away, we would reset. We now reset the wal to the prior index and use the truncate term and index.

Lastly if we receive a heartbeat with correct index but newer term just inherit.

Signed-off-by: Derek Collison <derek@nats.io>