Fix for situations where GC struggled to clean up things.
This mostly stemmed from subscriptions being held in per connection caches (routers and gateways). These held references to both clients and any icb, which usually were closures so would also hold reference to the struct target as well.
The code was not working correctly as it would only delete the subject in question, but when generation id changes we need to blow the whole cache similar to normal clients.

Signed-off-by: Derek Collison <derek@nats.io>