[FIXED/CHANGED] Reject server/cluster/gateway names with spaces
Having cluster name with spaces in the context of a leaf node could
cause problems when a subscription on the leaf node would be propagated
in the cluster's hub.

Since when a gateway is specified the name of the gateway needs to
match the name of the cluster, the restriction of "no spaces" is
added to gateway name.

Finally, in the case of the leafnode, if no cluster name is specified
the server name is used, so also applying the restriction to the
server name.

Resolves #5633

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>