[jsz] Add StreamLeaderOnly filter option
The stream state and replica info is only guaranteed to be accurate
when returned by the leader of a given stream. This new option
returns stream details only for the stream in which the server
is the leader for that stream. For systems with many streams this
can significantly reduce the amount of data returned when scraping
across all servers since non-leader details will likely be ignored.

Fix #5698

Signed-off-by: Byron Ruth <byron@nats.io>