[FIXED] JetStream: sending reply to Ack before processed (standalone)
The server would send the reply back to the client before removing
message from the stream, which could cause some issues if the
client expected the message to be removed after the AckSync() call
returns.
For instance, a WorkQueue that would have a max message of 1 per
subject and discard new message would fail a client that got a
message, ack it, and re-puplish this message. Depending on timing,
the client may get and error `maximum messages per subject exceeded`.

Used @pcsegal description of the problem and example code to write
a test that demonstrated the issue before the fix.

Resolves #5702

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>