(2.11) Internal: Intra-Process Queue fixes/improvements.
Go says that when using pool we should use pointer to a slice, not
the slice itself to save on copy when getting/putting things back.

That's what I tried to do, but I believe that the gain was defeated
by the fact that I was storing it as a []T in the queue object, and
more importantly, was returning the address of the local variable
when putting back in the pool. The way it was used, worked, but was
dangerous if queues were to be used differently. For instance with
implementation before this change, this loop would fail:
```
var elts []int
for i:=0;i<1000;i++{
    q.push(i+1)
    expected += i+1
    elts = q.pop()
    for _, v := range elts {
        sum += v
    }
    q.recycle(&elts)

    q.push(i+2)
    expected += i+2
    elts = q.pop()

    q.push(i+3)
    expected += i+3

    for _, v := range elts {
        sum += v
    }
    q.recycle(&elts)

    elts = q.pop()
    for _, v := range elts {
        sum += v
    }
    q.recycle(&elts)
}
if sum != expected {
    // ERROR!
}
```
If we use different variables, such as `elts1 := q.pop()`, etc.. then
it works. And again, the way it was used before did not cause issues.

In this PR, I am using []T and no sync.Pool. Instead, we will store up
to 5 slices in queue's specific "pool". In the server, I believe that
almost all use-cases are 1P-1C or NP-1C, so even a single pooled slice
may be enough, but since this is a generic IPQueue, we use this small
pool.

The other changes are to use the "in progress" count (and now size) when
checking for the limits, which we were not. So after a `pop()`, since the
queue is empty, the `push()` side would be able to store up to the limits
while the receiver was processing the popped elements.

I have added APIs to indicate progress when processing elements in the
"for-loop" that goes over the `pop()` result. I have modified code that
uses a queue with limit (only 2 so far) so that they use the new API.

I have added benchmarks so we can evaluate future changes. Aside from
the modifications to queue with limits, running the benchmark from
original code to new shows a slight improvement. Of course, updating
progress for each element popped() is slower than doing as a bulk, but
it allows for fine-grained control on the queue limits. And when using
with processing of JetStream messages, it is likely that the effect is
not relevant.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>