[FIXED] LeafNode: Don't report cluster name in varz/banner when none defined.
In situation where a Leaf server has no cluster specified, we internally
use the server name as the cluster name. We may need it in the protocol
so that the hub can suppress some messages to avoid duplicates.

We were however still reporting a cluster name in `/varz` and in the
banner on startup. This PR fixes both.

Resolves #5913

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>