When checking interest state for interest or workqueue streams, we would check all msgs from the streams first sequence through ack floor and up to delivered.
We do this to make sure our ack state is correct. In cases where there were alot of messages still in the stream due to offline or clow consumers, this could be a heavy load on a server.

This improvement uses LoadNextMsg() to efficiently skip ahead and we now remember our checked floor and do not repeat checks for messages below our check floor on subsequent runs.

Signed-off-by: Derek Collison <derek@nats.io>