Move calculating source consumer sequences back to inline.
We had added jitter to this however this happens when a stream becomes a leader, which will not be on startup and will already be spread out.
However if the stream is an R1, this delays this processing to after the client port is open whereas inline processes it before.

The reason this could be important is that the stream read lock is held while we reverse scan the stream looking for sequences, if the client port is open this can stall apps.

Signed-off-by: Derek Collison <derek@nats.io>