[FIXED] Websocket: handling of close message frame without status
The server was incorrectly sending back a close message frame with
status 1005 when not receiving a status in the close message frame
from the client. This was wrong. This status was to be used internally
to keep track that no status was received and therefore the server
should not send a status back.

Also, the server was sending status 1006 in some error conditions,
while the spec also states that this status must not be used to
set the status, instead it was to internally keep track that there
was an error condition.

Resolves #6259

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>