[ADDED] Option that disables fast producer stalling (drops msg instead)
Normally, when a producer detects that one of the consumer of a message
is falling behind, it will stall. Which means that if a message has
2 consumers and the first is "slow", then it will affect the timely
delivery to the second consumer.

With the new option `no_fast_producer_stall=true`, the server will
simply drop a message destined to a consumer that would have caused
the producer to stall. The message is still delivered to consumers
that are not falling behind.

The option can be config-reload'ed and if a message is dropped
due to fast-producer/slow-consumer, and the message was traced
(with deliver option), then the message trace egress event will
have an error indicating the reason why the message was not
delivered.

Resolves #6413

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>