(2.11) [ADDED] Monitoring: Include subscriptions in `Gatewayz`
When checking the `gatewayz` monitoring endpoint, if the user
requests accounts `?accs=1` or a specific account `?acc_name=B`
then we can now include the subscription list for each account
with `&subs=1` or the subscription details list with `&subs=detail`.

The subscription information will be contained in each account in
the `accounts` list of the `outbound_gateways` block.

Resolves #6473

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>