Optimize removeSeqPerSubject() for MaxMsgPerSubject == 1
This brings back an optimization that was removed with commit
79b41b122511df.

Without this optimization the first sequence number is lost from
the per-subject state whenever a message is added and the per
subject limit is enforced. This is happening a lot on a stream
with MaxMsgPerSubject set to 1. And it can lead to a dramatic
slowdown when catching up on a stream with sequence numbers spread
out wide.

Signed-off-by: Sven Neumann <sven.neumann@holoplot.com>