[FIXED] JetStream: cross-account mirror with subject transform issue
Using import/export, if one tries to limit the consumer create API
with specific stream name and subject filter, the sync of a mirror
would fail if using a single subject transform.

This PR checks that if there is no filter subject, but a single
entry in FilterSubjects, then we switch to simple FilterSubject
as to allow the use of the extended consumer create API that
allows for fine grained permissions.

Resolves #6751

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>