[monitoring] Fix jsz offset bug
The core issue was that the account names were not being
sorted unless offset was non-zero. However, the first page
will always be an offset of zero and thus there is no
determinism of the order.

Now the accounts are sorted and the logic is simplified
to handle both single account and multiple account queries.

Signed-off-by: Byron Ruth <byron@nats.io>