[FIXED] Add total to JSZ and revert behavior
In #6794, a limit default was enforced. However there was no
indication of size in the response. The new behavior broke
clients (Surveyor) that relied on getting everything in one
request.

This change reverts the behavior to allow for everything
to be returned in one request (fine for small deployments),
but it adds a Total field on the response so clients can
properly determine if subsequent requests are necessary.

Signed-off-by: Byron Ruth <byron@nats.io>