Print `Trapped signal` as Notice level message (#6800)
It would be useful to figure out what led to a change in the server
state post-factum.

## Test plan:
`nats-server --signal term`
and observer in the console:
```
[390785] 2025/04/16 16:19:23.666627 [INF] Server is ready
[390785] 2025/04/16 16:19:54.341025 [INF] Trapped "terminated" signal
[390785] 2025/04/16 16:19:54.341190 [INF] Initiating Shutdown...
[390785] 2025/04/16 16:19:54.341316 [INF] Server Exiting..       
```

There is https://pkg.go.dev/golang.org/x/sys/unix#SignalName, that we
could use to print names like SIGTERM here, but I guess we do not add
the dependency just for that.

Signed-off-by: Alex Bozhenko <alex@synadia.com>

Signed-off-by: Alex Bozhenko <alexbozhenko@gmail.com>