Optimise `enforceMsgPerSubjectLimit`
When all of the state was lost and we need to rebuild and enforce a max
messages per subject limit, then the `enforceMsgPerSubjectLimit` function
could linearly scan multiple times and become very expensive if there are
lots of subjects. This refactors the function to use new subject tree
intersection logic to more quickly work out if there is any overlap between
the subjects needing attention and the subjects in a given block, meaning
we only need to scan the store once.

Signed-off-by: Neil Twigg <neil@nats.io>