Speed up the creation of sync consumers for mirrors or sources after an extended downtime across a leafnode.
When a consumer for a source or a mirror is failing to be created, we backoff creation attempts.
If the failures were due to a downed leafnode, meaning the mirror or source is across a leafnode connection, the resync could take more time then desired after the leafnode reconnects.

This improves resync time by hooking into the leafnode's reconnect logic (either via connect or async info).
Once we detect the reconnect, we search for streams that are leaders, and are a mirror or sourcing from another stream, and do not have an active sync consumer.
If we detect this we will reset the consumer backoff and retry with just a small jitter backoff.

Signed-off-by: Derek Collison <derek@nats.io>