NRG: Don't compact on catchup until snapshot written
The `n.wal.Compact()` call before `n.installSnapshot()` is potentially
dangerous in a crash scenario as we could blow away the log without
having written the snapshot to disk yet. Once the snapshot is written,
`n.installSnapshot()` will compact the log itself to the same index (as
`lastIndex == n.pindex`).

Signed-off-by: Neil Twigg <neil@nats.io>