Change extended DataStats fields to pointers and export
This was introduced in #6967 to support gateway, route, and leaf
stats. However the DataStats type is used for client connection
stats (e.g. disconnect event) for which the additional fields
are not relevant. Having a pointer ensures the omitempty JSON
tag is respected and those nil fields are not serialized.

Signed-off-by: Byron Ruth <byron@nats.io>