Raft batching: Avoid sending leftovers right away
This commit removes a "pathological" case from the current Raft
batching mechanism: if the proposal queue contains more entries
than one batch can fit, then raft will send a full batch, followed
by a small batch containing the leftovers.
However, it was observed that its quite possible that while the
first batch was being stored and sent, clients may already have
pushed more stuff into the proposal queue in the meantime.
With this fix the server will compose and send a full batch, then
the leftovers are handled as follows: if more proposals were pushed
into the proposal queue, then we carry over the leftovers to the
next iteration. So that the leftovers are batched together with the
proposals that were added pushed in the meantime.
If there are no more proposals, then we send the leftovers right away.

For performance testing only at point.