Raft batching: Make stream writes async with sync_interval = always
This is the baseline for perfomance testing Raft's batching
capabilities. The behavior of the batching mechanism Raft
is easier to observe if disk writes are synchronous.
I.e we want to write() + fsync() the Raft log. So that
producers can easily keep the proposal queue busy.
To do so one can set "sync_interval= always". However, that
results in disastrous performance: when the leader receives
acks for a "big" batch of log entries, the upper layer will
write() and fsync() all entries in the batch, individually.

So this commit disables "sync always" on stream writes.
This *should* work in principle because the data is already in
the raft log. Alternatively, one could implement "group commit"
for streams, i.e. fsync() only one time after processing a batch
of entries.

For performance testing only at this point.