Raft batching: Reduce lock contention on raft mutex
This is an attempt to reduce contention between Propose() and
sendAppendEntry(). Change Propose() to acquire a read lock on Raft, and
avoid locking Raft during storeToWAL() (which potentially does IO and
may take a long time). This works as long as sendAppendEntry() is called
from the Raft's goroutine only, unless the entry does not require to be
stored to the Raft log. So the rest of the changes are for enforcing the
above requirement:
  * Change EntryLeaderTransfer so that it is not store to the Raft log.
  * Push EntryPeerState and EntrySnapshot entries to the proposal queue.
  * Make sure EntrySnapshot entries skip the leader check, so make sure
    those are not batched together with other entries.

For performance testing only at this point.