[FIXED] Message Tracing: Hop header set properly per gateway
Setting the "hop" header for each gateway could cause header
corruption. This is now fixed.

A test dealing with gateway has been improved to include more than
one gateway, which would have demonstrated the issue. The test now
passes and ensures that the hop is different per gateway.

Related to #7442

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>