[FIXED] LeafNode proxy validation
Saw the test TestConfigCheck flap. The reason is that the proxy
validation was done when parsing the "proxy" field, but there was
no guarantee that this field would be parsed after "urls", which
may lead to validation being successful (because there was no
URLs added yet). Move the check after all fields have been parsed
for a given remote.

Relates to #7242

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>