Introduce asynchronous file writing for stream snapshots
The monitorStream blocks for a potentially long time when creating and
installing a snapshot of the stream's state. This leads to increased
tail latency, especially under high load.
This change splits Raft's snapshotting mechanism in two steps. The first
step takes care of creating the snapshot file on disk, the second step
installs the snapshot. The first step uses a separate go routine to do
the writing. Thus removing the most expensive part off the main thread.
The RaftNode interface is extended with two methods:
`PrepareSnapshot` and `InstallPreparedSnapshot` to support this
this workflow.

Signed-off-by: Daniele Sciascia <daniele@nats.io>