Take snapshot creation off the leader's critical path
This is an experiment to allow a leader to request a snapshot
from a follower.
The idea is that the leader can, during normal operation,
delegate the task of creating snapshots to the followers.
This makes sense under the following assumptions:
the snapshot is relatively small, but expensive to compute.
Requesting a a ready made snapshot from a follower, offloads
this task from leader and avoids blocking leader.

This implementation is not complete, only for experimentation.