[FIXED] MQTT: Timing out waiting to load x retained messages
This could happen even if the server restored all retained messages.
The issue was caused by the fact that the check to see if all were
restored was based on the message sequence matching the last sequence
of the retained message stream, as opposed to the number of messages
in the stream.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>