[FIXED] MQTT: Retained message body possibly corrupt in cluster mode
When running in cluster mode, if a MQTT producer connects to server
S-1 and sends a retained message on "foo", other servers in the cluster
(say S-2) will be aware by having a consumer on the retained message
stream.
If an MQTT subscription matching the retained message is started
on S-2, that will make the retained message loaded into the cache.

If another retained message on that same subject is sent from S-1,
then when consuming the message from the retained message stream,
S-2 would replace the original retained message on "foo" by this
new message, but the body would at the very least contain extra
CR_LF ("\r\n") but could also contain some random data. This is
because the message body was not properly cloned and would still
reference an internal buffer (that may be reused).

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>