[FIXED] Websocket: limit buffer size during decompression of a frame
When the server would decompress a compressed websocket frame, it would
not limit the resulting size of the uncompressed buffer. Once uncompressed
the maximum payload size would still be used to reject messages that
are too big, but the server would have already uncompressed a possibly
very big buffer (if the frame contained highly compressed data).

This PR limits the number of bytes that are being decompressed using
the maximum payload size as a limit.

Credit goes to:
Pavel Kohout, Aisle Research (www.aisle.com) for reporting the issue
and providing a path.

The propose patched as been updated a bit (need to use atomic to
use the connection's max payload value) and some tweaks around
the use of the `io.LimitedReader`.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>