NRG: Removed leader may reappear in membership
When a leader is peer-removed it will send a EntryPeerRemove
followed by EntryLeaderTransfer. Normally a follower will
commit EntryPeerRemove after receiving processing EntryLeaderTranfer.
However, if the leader sneaks in a hearbeat in between, the
heartbeat will commit the EntryPeerRemove. When EntryPeerRemove
is committed, then `n.peers` will no longer contain the remove leader.
However, when EntryLeaderTransfer is processed, the follower would
re-create an entry for the removed leader in its `n.peers` set.

Signed-off-by: Daniele Sciascia <daniele@nats.io>