NRG: Decouple Raft transport layer [skip ci]
The Raft implementation was tightly coupled to the server's
internal client and send queue for the RPC communication.
This makes it difficult to test scenarios like network
partitions in deterministic manner.
The primary benefit of this change is improved testability. A
new mockTransport is introduced for testing, which allows for
simulating network partitions and for injecting behavior after
a message is sent.

Signed-off-by: Daniele Sciascia <daniele@nats.io>