NRG: Disjoint majorities during membership changes and network partitions
This commit fixes the following bugs:

- Inconsistent Cluster Size: When a leader was partitioned from the
  cluster, immediately after proposing a EntryAddPeer. The remaining
  nodes could end up with different view of the cluster size and
  quorum. So followers could have cluster size and would not match
  the number of peers in the peer set. A subsequent leader election,
  electing one of the followers, could break the quorum system.

- Incorrect Leader Election: It was possible for a new leader to be
  elected without a proper quorum. This could happen if a partition
  occurred after a new peer was proposed but before that change was
  committed. A follower could add the uncommitted peer to its peer
  set but would not update its cluster size and quorum, leading to
  an invalid election.

Both issues are solved by making sure that when a peer is added or
removed from the membership, the cluster size and quorum are adjusted
accordingly, at the same time. Followers would first add peers when
receiving the EntryAddPeer, and then adjusting the cluster size
only after commit. This patch changes this behavior such that the
cluster size and quorum are recomputed upon receiving the EntryAddPeer
/ EntryRemovePeer proposals. This is inline with the membership
protocol proposed in Ongaro's dissertation, section 4.1.
This patch also removes the concept of a "known" peer from the Raft
layer. A node would add a peer to its peer set when first receiving
the corresponding appendEntry, and on commit it would be marked as
"known". This distinction no longer applies.

Signed-off-by: Daniele Sciascia <daniele@nats.io>